requirements.txt
google-generativeai
gemini-api
gemini
python-dotenv
langchain
langchain-experimental
# streamlit

